projects/05-ctdmate/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.118.0
fastembed==0.7.3
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.77
langchain-openai==0.3.34
langchain-upstage==0.7.3
langsmith==0.4.31
loguru==0.7.3
markdownify==1.2.0
mmh3==5.2.0
mpmath==1.3.0
numpy==2.3.3
onnxruntime==1.23.0
openai==2.0.0
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.32.1
py_rust_stemmers==0.1.5
pydantic==2.11.9
pydantic_core==2.33.2
pypdf==4.3.1
python-dotenv==1.1.1
PyYAML==6.0.3
qdrant-client==1.15.1
rank-bm25==0.2.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.20.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
wheel==0.45.1
zstandard==0.25.0
